AI/requirements.txt
absl-py==1.4.0
asgiref==3.6.0
asttokens==2.2.1
astunparse==1.6.3
backcall==0.2.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
comm==0.1.3
debugpy==1.6.7
decorator==5.1.1
Django==3.2.13
django-environ==0.10.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
executing==1.2.0
filelock==3.12.0
flatbuffers==23.3.3
gast==0.4.0
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.54.0
h5py==3.8.0
idna==3.4
ipykernel==6.22.0
ipython==8.13.1
jax==0.4.8
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==8.2.0
jupyter_core==5.3.0
keras==2.12.0
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
ml-dtypes==0.1.0
mpmath==1.3.0
mysqlclient==2.1.1
nest-asyncio==1.5.6
networkx==3.1
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
parso==0.8.3
pickleshare==0.7.5
platformdirs==3.5.0
prompt-toolkit==3.0.38
protobuf==4.22.3
psutil==5.9.5
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.15.1
python-dateutil==2.8.2
pytz==2023.3
pywin32==306
PyYAML==6.0
pyzmq==25.0.2
requests==2.29.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
sympy==1.11.1
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
torch==2.0.0
tornado==6.3.1
traitlets==5.9.0
typing_extensions==4.5.0
urllib3==1.26.15
wcwidth==0.2.6
Werkzeug==2.3.3
wrapt==1.14.1
asgiref==3.6.0
Django==4.2.1
django-dotenv==1.4.2
environ==1.0
libclang==16.0.0
mpmath==1.3.0
networkx==3.1
numpy==1.23.5
oauthlib==3.2.2
pyasn1==0.5.0
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
sympy==1.12
traitlets==5.9.0
typing_extensions==4.5.0
urllib3==1.26.15
